Finished changing everything in presentation layer to fit with the new domain layer

+++ src/main/java/org/example/persistence/IPersistenceProgram.java
@@ -12,7 +12,7 @@ public interface IPersistenceProgram
     List<ProgramEntity> importPrograms(List<ProgramEntity> programEntities);
     boolean deleteProgram(long programId);
     boolean updateProgram(ProgramEntity programEntity);
-    ProgramEntity getProgramById(int id);
+    ProgramEntity getProgramById(long id);
     List<ProgramEntity> getAllPrograms();
     List<ProgramEntity> getProgramsByProducer(UserEntity producer);
     List<ProgramEntity> getProgramsByCompany(CompanyEntity companyEntity);
